[breaking] params - moved AggregateType from restx.common package to restx.types package and allowing to define custom implementations through ServiceLoader
